requirements.txt
niconico.py==2.0.8
redis==5.0.8
pymongo==4.8.0
requests==2.32.3
